requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-docx==1.1.0
